refactor(ui): extract PlanReminderFormDialog from the panel (#1044)
The create/edit dialog now owns all nine form field states, editingId,
the submitPending single-flight owner, validation, and the close guard;
the panel keeps only list/runs/query state, per-action pending, and
refresh. The panel opens the dialog with a PlanReminderFormSeed and
remounts it per open (key={formNonce}), so fields initialize from the
seed exactly where the old open-handler setters put them. The shared
PlanReminderSelect moves to its own leaf module used by both surfaces.

Co-migrated contracts pin the same invariants across the seam:
plan-reminder-panel-contract (submit owner/close guard now asserted on
the dialog), command-palette (title-input marker), utility-primitives
(preset button variants), and the copy-hygiene/stub-vocabulary scans.

One deliberate simplification: the card-menu edit/duplicate items drop
their submitPending || guard — unreachable, since the menu lives behind
the modal dialog whenever a submit can be in flight.
