+++ apps/desktop/src/main/__tests__/plan-reminder-panel-contract.test.ts
@@ -11,26 +11,37 @@ function blockBetween(source: string, start: string, end: string): string {
 }
 
 describe('Plan Reminder panel async action contract', () => {
+  // Issue #1044: the create/edit form (all field state + the submit owner)
+  // moved into PlanReminderFormDialog; the panel keeps list/runs/query state
+  // plus the per-action pending + refresh owners. Each invariant below is
+  // asserted against the component that now owns it.
   it('gates form submit and refresh before React commits disabled state', async () => {
     const ui = await readFile(resolve(REPO_ROOT, 'packages/ui/src/plan-reminder-panel.tsx'), 'utf8');
+    const dialog = await readFile(resolve(REPO_ROOT, 'packages/ui/src/plan-reminder-form-dialog.tsx'), 'utf8');
     const panelBlock = extractFunctionBlock(ui, 'PlanReminderPanel');
-    const submitBlock = blockBetween(panelBlock, 'async function submit', 'async function runPlanReminderAction');
+    const dialogBlock = extractFunctionBlock(dialog, 'PlanReminderFormDialog');
+    const submitBlock = blockBetween(dialogBlock, 'async function submit', 'return \\(');
     const refreshBlock = blockBetween(panelBlock, 'async function refreshFromPanel', 'return \\(');
 
-    assert.match(panelBlock, /const \[submitPending, setSubmitPending\] = useState\(false\)/);
+    assert.match(dialogBlock, /const \[submitPending, setSubmitPending\] = useState\(false\)/);
     assert.match(panelBlock, /const \[refreshPending, setRefreshPending\] = useState\(false\)/);
-    assert.match(panelBlock, /const submitPendingRef = useRef\(false\)/);
+    assert.match(dialogBlock, /const submitPendingRef = useRef\(false\)/);
     assert.match(panelBlock, /const refreshPendingRef = useRef\(false\)/);
     assert.match(
+      dialogBlock,
+      /return \(\) => \{\s*submitPendingRef\.current = false;\s*\};\s*\}, \[\]\)/,
+      'Plan Reminder pending form owner must be released when the dialog unmounts',
+    );
+    assert.match(
       panelBlock,
-      /return \(\) => \{\s*submitPendingRef\.current = false;\s*refreshPendingRef\.current = false;\s*pendingActionKeysRef\.current = new Set\(\);/,
-      'Plan Reminder pending form/refresh owners must be released when the panel unmounts',
+      /return \(\) => \{\s*refreshPendingRef\.current = false;\s*pendingActionKeysRef\.current = new Set\(\);/,
+      'Plan Reminder refresh/action pending owners must be released when the panel unmounts',
     );
 
     assert.match(
-      panelBlock,
-      /function closeReminderDialog\(\) \{\s*if \(submitPendingRef\.current\) return;\s*setFormDialogOpen\(false\);/,
-      'The form dialog must not close while a submit is still owned by the panel',
+      dialogBlock,
+      /function closeReminderDialog\(\) \{\s*if \(submitPendingRef\.current\) return;\s*props\.onOpenChange\(false\);/,
+      'The form dialog must not close while a submit is still owned by the dialog',
     );
     assert.match(
       submitBlock,
@@ -43,9 +54,9 @@ describe('Plan Reminder panel async action contract', () => {
       /finally \{\s*submitPendingRef\.current = false;\s*if \(planReminderMountedRef\.current\) setSubmitPending\(false\);/,
       'Plan Reminder submit owner must release without writing React state after unmount',
     );
-    assert.match(panelBlock, /const submitDisabled = !canCreate \|\| submitPending;/);
-    assert.match(panelBlock, /<form className="maka-plan-form" onSubmit=\{submit\} aria-busy=\{submitPending \? 'true' : undefined\}>/);
-    assert.match(panelBlock, /<UiButton type="submit" disabled=\{submitDisabled\}>/);
+    assert.match(dialogBlock, /const submitDisabled = !canCreate \|\| submitPending;/);
+    assert.match(dialogBlock, /<form className="maka-plan-form" onSubmit=\{submit\} aria-busy=\{submitPending \? 'true' : undefined\}>/);
+    assert.match(dialogBlock, /<UiButton type="submit" disabled=\{submitDisabled\}>/);
 
     assert.match(
       refreshBlock,
