+++ apps/desktop/src/main/__tests__/settings-roadmap-cleanup-contract.test.ts
@@ -352,6 +352,9 @@ describe('Settings coming-soon cleanup contract', () => {
         readRepo('packages/ui/src/skills-panel.tsx'),
         readRepo('packages/ui/src/daily-review-panel.tsx'),
         readRepo('packages/ui/src/plan-reminder-panel.tsx'),
+        // Issue #1044: keep the stub-vocabulary gate on the extracted
+        // plan-reminder form dialog (moved out of plan-reminder-panel.tsx).
+        readRepo('packages/ui/src/plan-reminder-form-dialog.tsx'),
         readRepo('packages/ui/src/relative-time.tsx'),
       ])
     ).join('\n');
